PlotterInterface API added to separate form Visualizer
