+++ .gitignore
@@ -18,20 +18,20 @@ test_autogalaxy/config/priors/autogalaxy.json
 test_autogalaxy/config/priors/autogalaxy.json
 test_autogalaxy/config/notation/.label.ini.swp
 test_autogalaxy/unit/output/
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/chi_squared_map.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/image.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/model_image.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/model_image_of_galaxy_0.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/model_image_of_galaxy_1.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/noise_map.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/normalized_residual_map.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/residual_map.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/signal_to_noise_map.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/subtracted_image_of_galaxy_0.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/fit_imaging/fits/subtracted_image_of_galaxy_1.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/inversion/fits/interpolated_errors.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/inversion/fits/interpolated_reconstruction.fits
-test_autogalaxy/unit/pipeline/files/plot/visualizer/inversion/fits/reconstructed_image.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/chi_squared_map.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/image.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/model_image.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/model_image_of_galaxy_0.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/model_image_of_galaxy_1.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/noise_map.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/normalized_residual_map.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/residual_map.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/signal_to_noise_map.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/subtracted_image_of_galaxy_0.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/fit_imaging/fits/subtracted_image_of_galaxy_1.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/inversion/fits/interpolated_errors.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/inversion/fits/interpolated_reconstruction.fits
+test_autogalaxy/unit/pipeline/files/plot/PlotterInterface/inversion/fits/reconstructed_image.fits
 test_autogalaxy/unit/data_temp/simulate_and_fit/image.fits
 test_autogalaxy/unit/data_temp/simulate_and_fit/noise_map.fits
 test_autogalaxy/unit/data_temp/simulate_and_fit/psf.fits
