+++ autogalaxy/imaging/model/visualizer.py
@@ -1,109 +1,94 @@
-from os import path
+import autofit as af
 
-from autogalaxy.imaging.fit_imaging import FitImaging
-from autogalaxy.imaging.plot.fit_imaging_plotters import FitImagingPlotter
-from autogalaxy.analysis.visualizer import Visualizer
+from autoarray import exc
 
-from autogalaxy.analysis.visualizer import plot_setting
+from autogalaxy.imaging.model.plotter_interface import PlotterInterfaceImaging
 
+class VisualizerImaging(af.Visualizer):
 
-class VisualizerImaging(Visualizer):
-    def visualize_fit_imaging(
-        self, fit: FitImaging, during_analysis: bool, subfolders: str = "fit_dataset"
+    def visualize_before_fit(
+            self,
+            analysis,
+            paths: af.AbstractPaths,
+            model: af.AbstractPriorModel,
     ):
         """
-        Visualizes a `FitImaging` object, which fits an imaging dataset.
+        PyAutoFit calls this function immediately before the non-linear search begins.
 
-        Images are output to the `image` folder of the `visualize_path` in a subfolder called `fit`. When
-        used with a non-linear search the `visualize_path` points to the search's results folder and this function
-        visualizes the maximum log likelihood `FitImaging` inferred by the search so far.
-
-        Visualization includes individual images of attributes of the `FitImaging` (e.g. the model data, residual map)
-        and a subplot of all `FitImaging`'s images on the same figure.
-
-        The images output by the `Visualizer` are customized using the file `config/visualize/plots.ini` under the
-        [fit] header.
+        It visualizes objects which do not change throughout the model fit like the dataset.
 
         Parameters
         ----------
-        fit
-            The maximum log likelihood `FitImaging` of the non-linear search which is used to plot the fit.
-        during_analysis
-            Whether visualization is performed during a non-linear search or once it is completed.
-        visuals_2d
-            An object containing attributes which may be plotted over the figure (e.g. the centres of mass and light
-            profiles).
+        paths
+            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
+            visualization and the pickled objects used by the aggregator output by this function.
+        model
+            The PyAutoFit model object, which includes model components representing the galaxies that are fitted to
+            the imaging data.
         """
 
-        def should_plot(name):
-            return plot_setting(section=["fit", "fit_imaging"], name=name)
+        plotter = PlotterInterfaceImaging(output_path=paths.image_path)
 
-        mat_plot_2d = self.mat_plot_2d_from(subfolders=subfolders)
+        plotter.imaging(dataset=analysis.dataset)
 
-        fit_plotter = FitImagingPlotter(
-            fit=fit, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
+        if analysis.adapt_images is not None:
+            plotter.adapt_images(adapt_images=analysis.adapt_images)
+
+    def visualize(
+        self,
+        analysis,
+        paths: af.DirectoryPaths,
+        instance: af.ModelInstance,
+        during_analysis: bool
+    ):
+        """
+        Output images of the maximum log likelihood model inferred by the model-fit. This function is called throughout
+        the non-linear search at regular intervals, and therefore provides on-the-fly visualization of how well the
+        model-fit is going.
+
+        The visualization performed by this function includes:
+
+        - Images of the best-fit galaxies, including the images of each of its galaxies.
+
+        - Images of the best-fit `FitImaging`, including the model-image, residuals and chi-squared of its fit to
+          the imaging data.
+
+        - The adapt-images of the model-fit showing how the galaxies are used to represent different galaxies in
+          the dataset.
+
+        The images output by this function are customized using the file `config/visualize/plots.yaml`.
+
+        Parameters
+        ----------
+        paths
+            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
+            visualization, and the pickled objects used by the aggregator output by this function.
+        instance
+            An instance of the model that is being fitted to the data by this analysis (whose parameters have been set
+            via a non-linear search).
+        during_analysis
+            If True the visualization is being performed midway through the non-linear search before it is finished,
+            which may change which images are output.
+        """
+        fit = analysis.fit_from(instance=instance)
+
+        plotter = PlotterInterfaceImaging(output_path=paths.image_path)
+        plotter.imaging(dataset=analysis.dataset)
+
+        try:
+            plotter.fit_imaging(fit=fit, during_analysis=during_analysis)
+        except exc.InversionException:
+            pass
+
+        galaxies = fit.galaxies_linear_light_profiles_to_light_profiles
+
+        plotter.galaxies(
+            galaxies=galaxies, grid=fit.grid, during_analysis=during_analysis
         )
-
-        fit_plotter.figures_2d(
-            data=should_plot("data"),
-            noise_map=should_plot("noise_map"),
-            signal_to_noise_map=should_plot("signal_to_noise_map"),
-            model_image=should_plot("model_data"),
-            residual_map=should_plot("residual_map"),
-            normalized_residual_map=should_plot("normalized_residual_map"),
-            chi_squared_map=should_plot("chi_squared_map"),
+        plotter.galaxies_1d(
+            galaxies=galaxies, grid=fit.grid, during_analysis=during_analysis
         )
-
-        fit_plotter.figures_2d_of_galaxies(
-            subtracted_image=should_plot("subtracted_images_of_galaxies"),
-            model_image=should_plot("model_images_of_galaxies"),
-        )
-
-        if should_plot("subplot_fit"):
-            fit_plotter.subplot_fit()
-
-        if should_plot("subplot_of_galaxies"):
-            fit_plotter.subplot_of_galaxies()
-
-        if not during_analysis and should_plot("all_at_end_png"):
-            mat_plot_2d = self.mat_plot_2d_from(subfolders=path.join(subfolders, "end"))
-
-            fit_plotter = FitImagingPlotter(
-                fit=fit, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
-            )
-
-            fit_plotter.figures_2d(
-                data=True,
-                noise_map=True,
-                signal_to_noise_map=True,
-                model_image=True,
-                residual_map=True,
-                normalized_residual_map=True,
-                chi_squared_map=True,
-            )
-
-            fit_plotter.figures_2d_of_galaxies(subtracted_image=True, model_image=True)
-
-        if not during_analysis and should_plot("all_at_end_fits"):
-            mat_plot_2d = self.mat_plot_2d_from(
-                subfolders=path.join(subfolders, "fits"), format="fits"
-            )
-
-            fit_plotter = FitImagingPlotter(
-                fit=fit, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
-            )
-
-            fit_plotter.figures_2d(
-                data=True,
-                noise_map=True,
-                signal_to_noise_map=True,
-                model_image=True,
-                residual_map=True,
-                normalized_residual_map=True,
-                chi_squared_map=True,
-            )
-
-            fit_plotter.figures_2d_of_galaxies(
-                subtracted_image=True,
-                model_image=True,
+        if fit.inversion is not None:
+            plotter.inversion(
+                inversion=fit.inversion, during_analysis=during_analysis
             )
