+++ test_autogalaxy/imaging/model/test_plotter_interface_imaging.py
@@ -3,14 +3,34 @@ from os import path
 import pytest
 
 import autogalaxy as ag
-from autogalaxy.imaging.model.visualizer import VisualizerImaging
+from autogalaxy.imaging.model.plotter_interface import PlotterInterfaceImaging
 
 directory = path.dirname(path.abspath(__file__))
 
 
 @pytest.fixture(name="plot_path")
-def make_visualizer_plotter_setup():
+def make_plotter_interface_plotter_setup():
     return path.join("{}".format(directory), "files")
+
+
+def test__visualizes_imaging__uses_configs(
+    imaging_7x7, include_2d_all, plot_path, plot_patch
+):
+    if path.exists(plot_path):
+        shutil.rmtree(plot_path)
+
+    plotter_interface = PlotterInterfaceImaging(output_path=plot_path)
+
+    plotter_interface.imaging(dataset=imaging_7x7)
+
+    assert path.join(plot_path, "subplot_dataset.png") in plot_patch.paths
+
+    plot_path = path.join(plot_path, "dataset")
+
+    assert path.join(plot_path, "data.png") in plot_patch.paths
+    assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
+    assert path.join(plot_path, "psf.png") in plot_patch.paths
+    assert path.join(plot_path, "signal_to_noise_map.png") not in plot_patch.paths
 
 
 def test__visualizes_fit_imaging__uses_configs(
@@ -23,9 +43,9 @@ def test__visualizes_fit_imaging__uses_configs(
     if path.exists(plot_path):
         shutil.rmtree(plot_path)
 
-    visualizer = VisualizerImaging(visualize_path=plot_path)
+    plotter_interface = PlotterInterfaceImaging(output_path=plot_path)
 
-    visualizer.visualize_fit_imaging(
+    plotter_interface.fit_imaging(
         fit=fit_imaging_x2_galaxy_inversion_7x7, during_analysis=False
     )
