+++ test_autogalaxy/interferometer/model/test_visualizer_interferometer.py
@@ -1,14 +1,30 @@
 from os import path
 import pytest
 
-from autogalaxy.interferometer.model.visualizer import VisualizerInterferometer
+from autogalaxy.interferometer.model.plotter_interface import PlotterInterfaceInterferometer
 
 directory = path.dirname(path.abspath(__file__))
 
 
 @pytest.fixture(name="plot_path")
-def make_visualizer_plotter_setup():
+def make_plotter_interface_plotter_setup():
     return path.join("{}".format(directory), "files")
+
+
+def test__visualizes_interferometer__uses_configs(
+    interferometer_7, include_2d_all, plot_path, plot_patch
+):
+    plotter_interface = PlotterInterfaceInterferometer(output_path=plot_path)
+
+    plotter_interface.interferometer(dataset=interferometer_7)
+
+    assert path.join(plot_path, "subplot_dataset.png") in plot_patch.paths
+
+    plot_path = path.join(plot_path, "dataset")
+
+    assert path.join(plot_path, "data.png") in plot_patch.paths
+    assert path.join(plot_path, "u_wavelengths.png") not in plot_patch.paths
+    assert path.join(plot_path, "v_wavelengths.png") not in plot_patch.paths
 
 
 def test__visualizes_fit_interferometer__uses_configs(
@@ -18,9 +34,9 @@ def test__visualizes_fit_interferometer__uses_configs(
     plot_path,
     plot_patch,
 ):
-    visualizer = VisualizerInterferometer(visualize_path=plot_path)
+    PlotterInterface = PlotterInterfaceInterferometer(output_path=plot_path)
 
-    visualizer.visualize_fit_interferometer(
+    PlotterInterface.fit_interferometer(
         fit=fit_interferometer_x2_galaxy_inversion_7x7, during_analysis=True
     )
